+++ lib/interfaces/error_interface.rb
@@ -10,9 +10,18 @@ module LowType
       @output_size = LowType.config.output_size
     end
 
+    def error_type
+      raise NotImplementedError
+    end
+
+    def error_message(value:)
+      raise NotImplementedError
+    end
+
     def output(value:)
       case @output_mode
       when :type
+        # TODO: Show full type structure in error output instead of just the type of the supertype.
         value.class
       when :value
         value.inspect[0...@output_size]
@@ -21,12 +30,16 @@ module LowType
       end
     end
 
-    def error_type
-      raise NotImplementedError
-    end
+    def backtrace(backtrace:, hidden_paths:)
+      # Remove LowType defined method file paths from the backtrace.
+      filtered_backtrace = backtrace.reject { |line| hidden_paths.find { |file_path| line.include?(file_path) } }
 
-    def error_message(value:)
-      raise NotImplementedError
+      # Add the proxied file to the backtrace.
+      proxy_file_backtrace = "#{file.path}:#{file.line}:in '#{file.scope}'"
+      from_prefix = filtered_backtrace.first.match(/\s+from /)
+      proxy_file_backtrace = "#{from_prefix}#{proxy_file_backtrace}" if from_prefix
+
+      [proxy_file_backtrace, *filtered_backtrace]
     end
   end
 end