+++ CHANGELOG.md
@@ -3,16 +3,24 @@
 All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
-Minor features that don't break backwards compatibiliy are released as patches.
+Minor features that don't break backwards compatibility are released as patches.
 
-## 1.1.0 [UNRELEASED]
+## 1.2.0 [UNRELEASED]
 
 ### Added
 
 - Support dynamic expressions in methods and return types at runtimem (like `type()` already does)
 - `Boolean` type support
-- Complex types support
+- Complex types validation
 - Error mode config
+
+## 1.1.0
+
+### Added
+
+- Deep type checking
+- Array subtype expressions `Array[String | nil]`
+- Complex types support
 
 ## 1.0.8